refactor: useSchedulModal 훅 분리
